+++ base_app/src/app/dataset-tables/title-bar-test.component.spec.ts
@@ -2,7 +2,7 @@ import { AppComponent } from '../app.component';
 import { AppTabNormalComponent } from '../app-tab-normal/app-tab-normal.component'
 import { AppTabOverlayComponent } from '../app-tab-overlay/app-tab-overlay.component'
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { CommandButtonComponent } from '../cmdsets/command-button';
+import { CommandButtonComponent } from '../command-button/command-button';
 import { CUSTOM_ELEMENTS_SCHEMA, DebugElement } from '@angular/core';
 import { ImageOverlaysComponent } from '../app-tab-overlay/image-overlays';
 import { OverlayCmdBarComponent } from '../app-tab-overlay/overlay-cmd-bar';
@@ -10,7 +10,7 @@ import { OverlayPageComponent } from '../app-tab-overlay/overlay-page';
 import { DatasetTableComponent } from './dataset-table';
 import { PortalModule } from '@angular/cdk/portal';
 import { MatTabsModule } from '@angular/material/tabs';
-import { PropDefinedTableComponent } from './prop-defined-table';
+import { PropDefinedTableComponent } from '../prop-defined-table/prop-defined-table';
 import {CommonModule} from '@angular/common';
